add new not work fun

+++ app/src/main/java/com/example/covid19_vaccine/data/domain/VaccineData.kt
@@ -7,7 +7,7 @@ data class VaccineData(
     val date: String,
     val people_vaccinated: BigInteger?,
     val people_fully_vaccinated: BigInteger?,
-    val daily_vaccinations: BigInteger?,
+    val daily_vaccinations: Int?,
     val people_vaccinated_per_hundred: Double?,
     val people_fully_vaccinated_per_hundred: Double?,
     val daily_vaccinations_per_million: Int?
